+++ cpskin/menu/upgrades.py
@@ -11,3 +11,8 @@ def upgrade_1001_to_1002(context):
 
 def upgrade_1002_to_1003(context):
     context.runImportStepFromProfile(PROFILE_ID, 'jsregistry')
+
+
+def move_cpskin_actions(context):
+    context.runImportStepFromProfile('profile-cpskin.menu:to1005', 'actions')
+    context.runImportStepFromProfile('profile-cpskin.menu:default', 'actions')